+++ src/main/java/com/blackrook/json/annotation/JSONCollectionType.java
@@ -22,6 +22,7 @@ import com.blackrook.json.JSONObject;
  *     <li><b>Collection&lt;T&gt;</b> - ArrayList&lt;T&gt;</li>
  * </ul>
  * If not a Collection, a {@link JSONConversionException} will occur.
+ * @since [NOW]
  */
 @Target({ElementType.METHOD, ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)