+++ docs/CHANGELOG.md
@@ -3,6 +3,12 @@ JSON (C) Black Rook Software 2019
 by Matt Tropiano et al. (see AUTHORS.txt)
 
 
+Changed in [NOW]
+----------------
+
+- Added Generic collection hints for erased types (JSONCollectionType, JSONMapType).
+
+
 Changed in 1.0.0
 ----------------